Move metadata to its own subproject
