+++ metadata/interface.ts
