+++ metadata/tsconfig.json
@@ -1,6 +1,5 @@
 {
   "extends": "../tsconfig-base.json",
   "compilerOptions": {},
-  "references": [{ "path": "../localization" }, { "path": "../metadata" }],
   "include": ["**/*"]
 }
