+++ src/plugins/intellisense/index.tsx
@@ -13,7 +13,7 @@ import { DCGView, MountedComponent, unmountFromNode } from "#DCGView";
 import { MathQuillField, MathQuillView } from "#components";
 import { ItemModel, TextModel, Calc } from "#globals";
 import { PluginController } from "#plugins/PluginController.ts";
-import { getMetadata } from "#plugins/manage-metadata/manage.ts";
+import { getMetadata } from "#plugins/manage-metadata/sync.ts";
 import { hookIntoOverrideKeystroke } from "#utils/listenerHelpers.ts";
 import { isDescendant } from "#utils/utils.ts";
