+++ src/plugins/intellisense/state.tsx
@@ -4,7 +4,7 @@ import { getTextModeConfig } from "../text-mode";
 import { mapAugAST } from "./latex-parsing";
 import { Calc, ItemModel } from "#globals";
 import { rootKeys } from "#plugins/find-replace/backend.ts";
-import Metadata from "#plugins/manage-metadata/interface.ts";
+import Metadata from "metadata/interface";
 import { get } from "#utils/utils.ts";
 
 function getOrMakeKey<K, V>(map: Map<K, V>, k: K, v: () => V) {
