+++ src/plugins/manage-metadata/sync.ts
@@ -1,6 +1,6 @@
-/* eslint-disable @typescript-eslint/no-dynamic-delete */
-import Metadata, { Expression } from "./interface";
-import migrateToLatest from "./migrate";
+import Metadata from "#metadata/interface.ts";
+import migrateToLatest from "#metadata/migrate.ts";
+import { getBlankMetadata, isBlankMetadata } from "#metadata/manage.ts";
 import { Calc, Console, FolderModel, TextModel } from "#globals";
 import { List } from "#utils/depUtils.ts";
 
@@ -68,63 +68,12 @@ export function setMetadata(metadata: Metadata) {
   }
 }
 
-export function getBlankMetadata(): Metadata {
-  return {
-    version: 2,
-    expressions: {},
-  };
-}
-
-export function isBlankMetadata(metadata: Metadata) {
-  return (
-    Object.keys(metadata.expressions).length === 0 &&
-    Object.keys(metadata).length === 2
-  );
-}
-
 function cleanMetadata(metadata: Metadata) {
   /* Mutates metadata by removing expressions that no longer exist */
   for (const id in metadata.expressions) {
     if (Calc.controller.getItemModel(id) === undefined) {
+      // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
       delete metadata.expressions[id];
-    }
-  }
-}
-
-export function changeExprInMetadata(
-  metadata: Metadata,
-  id: string,
-  obj: Partial<Expression>
-) {
-  /* Mutates metadata by spreading obj into metadata.expressions[id],
-  with default values deleted */
-  const changed = metadata.expressions[id] ?? {};
-  for (const _key in obj) {
-    const key = _key as keyof Expression;
-    const value = obj[key];
-    if (value !== getDefaultValue(key)) {
-      changed[key] = value;
-    } else {
-      delete changed[key];
-    }
-  }
-  if (Object.keys(changed).length === 0) {
-    delete metadata.expressions[id];
-  } else {
-    metadata.expressions[id] = changed;
-  }
-}
-
-function getDefaultValue(key: keyof Expression) {
-  switch (key) {
-    case "pinned":
-    case "errorHidden":
-    case "glesmos":
-    case "glesmosLinesConfirmed":
-      return false;
-    default: {
-      const _exhaustive: never = key;
-      return _exhaustive;
     }
   }
 }
