+++ src/plugins/text-mode/modify.ts
@@ -23,7 +23,7 @@ import { ChangeSpec, StateEffect } from "@codemirror/state";
 import { EditorView } from "@codemirror/view";
 import { GraphState, NonFolderState } from "@desmodder/graph-state";
 import { Calc, DispatchedEvent } from "#globals";
-import Metadata from "#plugins/manage-metadata/interface.ts";
+import Metadata from "metadata/interface";
 
 // @settings related
 const settingsEvents = [
