+++ tsconfig.json
@@ -2,6 +2,7 @@
   "include": ["./esbuild.mjs", "./.eslintrc.js"],
   "references": [
     { "path": "./scripts" },
+    { "path": "./metadata" },
     { "path": "./text-mode-core" },
     { "path": "./localization" },
     { "path": "./src" }
